requirements.txt
weaviate-client>=4.5.0
sentence-transformers>=2.2.2
numpy>=1.23.0
pillow>=9.0.0
matplotlib>=3.5.0
langchain>=0.1.0
langchain-groq>=0.1.0
requests>=2.28.0
python-dotenv>=1.0.0